chore: rename folder name
